backend/requirements.txt
fastapi[standard]
groq
python-dotenv
bandit
PyJWT
passlib[bcrypt]
sqlalchemy
